refactor: LottoNumberMaker 추가하여, numbers 유틸 클래스와 로또 번호 추출에 대한 책임 분리
