+++ src/test/java/dev/dahye/lotto/domain/LottoNumbersTest.java
@@ -37,11 +37,4 @@ class LottoNumbersTest {
         List<Integer> changedNumbers = LottoNumbers.getNumbers();
         assertThat(changedNumbers).isNotEmpty();
     }
-
-    @Test
-    @DisplayName("로또 값 객체가 가지고 있는 값보다 큰 값을 요청하는 경우 IllegalArgumentException throw")
-    void valid_lotto_numbers_max_value() {
-        assertThrows(IllegalArgumentException.class,
-                () -> LottoNumbers.createShuffled(LottoNumbers.MAX_VALUE + 1));
-    }
 }
